+++ features/support/app_pages/shop/admin/shop_admin_categories_page.rb
@@ -29,9 +29,23 @@ class ShopAdminCategories < IDmeBase
     js_accept
   end
 
+  def click_on_new_category
+    find("#DataTables_Table_0 tbody td:first-child").click
+  end
+
+  def update_new_category
+    fill_in("category_description", :with => "Getting super swole")
+    click_button "Update"
+  end
+
+  def click_cancel_button
+    click_link "Cancel"
+  end
+
   def submit_invalid_category
     click_button_for_new_category_page
     fill_in("category_label", :with =>"")
     click_button "Create"
   end
+  
 end